pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "q-materials-database"
version = "0.1.0"
description = ""
authors = [
    {name = "Mamba-Grant",email = "mumbo1002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "psycopg-binary (>=3.2.9,<4.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "jupyterlab (>=4.4.4,<5.0.0)",
    "ase (>=3.25.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
